Extract ABI related node extensions
