+++ crates/solidity-v2/outputs/cargo/ast/src/abi/mod.rs
@@ -1,3 +1,5 @@
+mod node_extensions;
+
 use std::cmp::Ordering;
 use std::rc::Rc;
