+++ crates/solidity-v2/outputs/cargo/ast/src/ast/node_extensions/mod.rs
@@ -1,10 +1,7 @@
 mod contracts;
 pub use contracts::ContractBase;
 
-mod errors;
-mod events;
 mod expressions;
 mod functions;
 mod identifiers;
 mod source_unit;
-mod state_variables;
