+++ crates/solidity-v2/outputs/cargo/ast/src/ast/node_extensions/source_unit.rs
@@ -1,5 +1,4 @@
 use super::super::{ContractDefinition, SourceUnitMember, SourceUnitStruct};
-use crate::abi;
 
 impl SourceUnitStruct {
     pub fn file_id(&self) -> String {
@@ -14,20 +13,6 @@ impl SourceUnitStruct {
                     Some(contract)
                 } else {
                     None
-                }
-            })
-            .collect()
-    }
-
-    pub fn compute_contracts_abi(&self) -> Vec<abi::ContractAbi> {
-        let file_id = self.file_id();
-        self.contracts()
-            .iter()
-            .filter_map(|contract| {
-                if contract.abstract_keyword() {
-                    None
-                } else {
-                    contract.compute_abi_with_file_id(file_id.clone())
                 }
             })
             .collect()
